+++ lib/settings.js
@@ -8,14 +8,12 @@ const path = require('path')
 const instantiate = require('./instantiate')
 const hotRequire = require('./hot-require')
 
-function load (filename) {
+function load (filename, overrides = {}) {
   // take a path + object, turn it into real objects
   const dirname = path.dirname(filename)
   const settings = Object.assign({
     NAME: 'spife',
     IS_EXTERNAL: true,
-    ENABLE_FORM_PARSING: false,
-    MAX_REQUEST_BODY_SIZE: 1 << 20,
     METRICS: null,
     REQUEST_ID_HEADERS: ['request-id'],
     ON_CLIENT_ERROR: () => {},
@@ -26,7 +24,7 @@ function load (filename) {
     NODE_ENV: process.env.NODE_ENV || 'development',
     DEBUG: false,
     HOT: false
-  }, require(filename))
+  }, require(filename), overrides)
 
   if (!settings.ROUTER) {
     throw new Error('spife settings files should specify a ROUTER value')
