+++ package.json
@@ -62,6 +62,7 @@
     "range-parser": "^1.2.0",
     "replify": "^1.2.0",
     "reverse": "^4.0.0",
+    "shot": "^4.0.5",
     "statuses": "^1.3.1",
     "table": "^4.0.1",
     "tacks": "^1.2.6",
